+++ @types/express/index.d.ts
@@ -1,0 +1,8 @@
+/// <reference types="aws-lambda" />
+
+declare namespace Express {
+  interface Request {
+    endpoint?: string
+    event?: AWSLambda.APIGatewayProxyEvent
+  }
+}